product listing updated

+++ database/migrations/2025_03_15_230733_create_products_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->decimal('price', 10, 2);
-            $table->decimal('sale_price', 10, 2)->nullable();
+            $table->integer('price');
+            $table->integer('sale_price')->nullable();
             $table->string('tags')->nullable(); // Comma-separated values
             $table->text('short_description')->nullable();
             $table->text('description')->nullable();
